requirements.txt
transformers>=4.45.0
torch>=2.1.0
accelerate>=0.34.0
huggingface_hub>=0.25.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
rdkit-pypi>=2023.9.1
tqdm>=4.65.0
tabulate>=0.9.0
